Add improved support for working with Retina images.
This fixes an important bug in doc generation where Retina filenames
weren't being generated when there were image files with the same name
in the tree. When Sphinx has two images in a tree sharing the same
filename (or one directory references another's image), one of the
images get renamed to include a numeric suffix. The reason is that all
images end up in the same directory and are not namespaced in any other
way. When this happened, we'd fail to create an equivalent Retina image,
breaking support on one or more of the pages.

This also adds a configuration option for Retina suffixes like `@2x`
and `@3x`. These are the default suffixes that will be tried, but
additional ones can be added on `conf.py`.

Testing Done:
Built docs for Review Board and saw that the Retina image issues we've
been encountering for one of the images has gone away.

Reviewed at https://reviews.reviewboard.org/r/9359/