Add configurable title support for the `:http:` role.
The `:http:` role has always required a numeric status code and returned
a fixed result in the form of "HTTP <code> <name>", but this isn't
always suitable for docs. This change does a few things to improve this.

`:http:` now takes an optional title using the same format used for
references, allowing the title to be changed per-instance.

The format string for the title can also be set in the Sphinx
configuration, allowing a different title to be used by default
throughout the docs.

This title can also be adjusted per-page or for particular sections in a
page by using the new `.. http-status-codes-format::` directive. When
called, the format string provided will be used for all instances of the
roll. This can be called again without an argument to go back to using
the default for the docs.

Testing Done:
Made use of every one of these new capabilities in a docs file, ensuring
that they all work in combination with each other.

Reviewed at https://reviews.reviewboard.org/r/9379/