Improve handling of multi-line argument types.
The code for handling multi-line types for arguments in "Args" sections
could lose spaces when breaking after a comma (for defining things like
"optional" or "unused" after a type). This happened because we stripped
each line before joining.

This change improves the processing, stripping only when part of an
argument reference (anything with a leading or trailing `.`).

Testing Done:
Tested with single-line argument definitions and multiline (with `.` on
the previous line, next line, both, and with `, optional` and other text.

Reviewed at https://reviews.reviewboard.org/r/9986/