Add unit tests for beanbag-docutils.
beanbag-docutils has a lot of complex code, and as we go it's getting
more difficult to make changes without risking regressions.

This change introduces the beginnings of a test suite, which runs
through all the main roles, directives, and Beanbag docstring parsing to
ensure that the output we get is correct now and in the future with new
versions of Sphinx.

Some things are not tested. For instance, the image collection and
Retina handling. These are a bit more complex, and are going to require
further work before unit tests can be added.

There are a few fixes found in the unit test runs, as well as a few
small fixes to typos in docstrings and a URL tweak for Wikipedia.

Testing Done:
Unit tests passed with Sphinx 1.8.5 (last version to support Python 2.7)
and 3.2.1 (latest public version).

Reviewed at https://reviews.reviewboard.org/r/11140/