Update the intersphinx_utils extension for Sphinx 3+
In Sphinx 3 and up, the listeners for missing references are represented
as a list. Earlier versions use a dictionary. This change updates the
code to check for either form, bringing compatibility to both.

Testing Done:
Been using this to build docs on Sphinx 2 and 3.

Reviewed at https://reviews.reviewboard.org/r/11485/