Fix lists in version information sections.
When using `Version Added`, `Version Changed`, or `Deprecated` with a
numeric or bullet list, the rendering of the lines would break. This
wasn't anything special with these sections, and in fact likely applied
to others as well. Instead, it was a mistake make in the re-indent logic
for putting content into a section.

The logic was supposed to dedent a block of lines, removing the common
leading whitespace, and to then re-indent with 3 spaces. The `3`
parameter for `_indent()` was put in the wrong place, and this instead
was being passed to `_dedent()`, taking the place of a parameter that,
if truthy, would dedent all whitespace.

This led to subsequent lines in multi-line list items turning into new
paragraphs.

This change fixes that argument issue.

Testing Done:
Unit tests pass.

Saw that this fixed some documentation I was working on.

Reviewed at https://reviews.reviewboard.org/r/12513/