Add support a Type section.
This formalizes our `Type` section, which we use for properties and
attributes. It properly links the type, and even allows for an optional
description to help formally describe the type (using `Tuple` or `Keys`
or similar).

Testing Done:
Unit tests pass.

Tested this with our documentation. Saw that the types were properly set
up as sections like "Returns" and had links to types.

Reviewed at https://reviews.reviewboard.org/r/12517/