Fix autodoc type alias injection to fix incremental doc builds.
Our `autodoc_utils` Sphinx extension tries to determine some built-in
type mappings to feed to napoleon. In newer versions of Sphinx,
napoleon's mappings get set again in the autodoc's version of the
setting, but too late for us. This ultimately results in our injected
settings always appearing "new" compared to the stored settings for the
doc build, and that forces a full rebuild every time docs are built.

We now set this for both napoleon and autodoc, ensuring this does not
happen.

Testing Done:
Successfully built docs incrementally on the versions of Sphinx we test
with.

Reviewed at https://reviews.reviewboard.org/r/12738/