Fix GitHub code link generation when failing to resolve tracking branches.
When failing to resolve tracking branches, we'd end up with a `None`
result, which would then get fed back into a `git rev-parse`, causing an
error.

This change avoids this situation, and ensures we don't try to generate
a link if we fail to locate a branch (since it won't have a stable SHA
to point to). It also and asserts that we don't pass in either a `None`
for the command or any falsey values in the list.

Testing Done:
Unit tests pass.

Successfully built Review Board docs.

Reviewed at https://reviews.reviewboard.org/r/12739/