Exclude __annotations__ by default from docs.
This adds `__annotations__` to the list of attributes to exclude on
modules, classes, and functions. These were noisy and not beneficial for
documentation, as the information from annotations are already pulled
out for use.

Testing Done:
Built docs and verified that the `__annotations__` attributes weren't
being shown.

Reviewed at https://reviews.reviewboard.org/r/12740/