Add a missing dependency on typing-extensions.
Recent work on autodoc-utils added an import for `typing-extensions`,
but this wasn't added to the list of dependencies. That ended up
breaking builds.

This change adds that dependency, using 4.6 as the minimum version. This
is pretty old, but is the first version to introduce the
`__orig_bases__` feature we're using. Most installs will have a newer
version, but this keeps the constraints from being too tight and causing
upgrades unnecessarily.

Testing Done:
Successfully tested in a new virtualenv.

Reviewed at https://reviews.reviewboard.org/r/14541/